migration fix

+++ db/migrate/20160425195844_remove_scheduler_runs.rb
@@ -1,6 +1,8 @@
 class RemoveSchedulerRuns < ActiveRecord::Migration
   def up
-    drop_table :scheduler_runs
+    if ActiveRecord::Base.connection.table_exists? 'scheduler_runs'
+      drop_table :scheduler_runs
+    end
   end
 
   def down
